Add files via upload
This code generates graphs to check for normality of the optimal allocation ratios. It stores the optimal allocation ratios for different (300,400,500,600) sample sizes, and check the normality of the optimal allocation ratios for each of the sample size. The Monte Carlo simulation runs for 5000 times. 
For each of sample size, we generate a total of 5000 optimal allocation ratio values for each of the three optimal allocation ratios. Then we generate the Quantile Quantile (Q-Q) plot for each of the optimal allocation ratios. This Q-Q plot further helps us to understand for which sample size the normality assumptions is possible.